try to remove version option in case we are uninstalling a dimension and remove a log_conversion column in a VisitDimension only if it implements a conversion

+++ core/Plugin/Dimension/VisitDimension.php
@@ -159,6 +159,10 @@ abstract class VisitDimension extends Dimension
         }
 
         try {
+            if (!$this->isHandlingLogConversion()) {
+                return;
+            }
+
             $sql = "ALTER TABLE `" . Common::prefixTable('log_conversion') . "` DROP COLUMN `$this->columnName`";
             Db::exec($sql);
         } catch (Exception $e) {
